14 Jul 2010

+++ vid_gl/r_mesh.h
@@ -73,7 +73,7 @@ typedef struct
 		int	lastPoly;
 		uint	dlightbits;
 		uint	modhandle;
-		float	placeTime;	// decal createtime
+		uint	placeTime;	// decal createtime
 	};
 	uint		shadowbits;
 } meshbuffer_t;
